Move module jsdoc tests to module tests
